requirements.txt
pyspark>=3.5.3
s3fs>=2024.10.0
python-dotenv>=1.0.1
